requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
pillow==10.0.0
pdf2image==1.16.3
reportlab==4.0.4
python-dotenv==1.0.0
mysql-connector-python==8.0.33
